refactor(browse): split the client and clear the lint warnings
`biome check` reported 24 warnings and one formatter error against the files
this branch adds. They were the same rules `finding.ts` and
`finding.test.ts` already trip, but the PR checklist asks that changes
generate no new warnings, and inheriting a rule violation is not a reason to
keep adding to it.

The browse client is split along the seams it already had, following #138
where the marketing and inventory god modules were broken up the same way:

  browseTypes.ts       shared interfaces and the sort enum
  browseFilter.ts      the Browse filter grammar and its conflict check
  browseMappers.ts     pure response mappers
  browseValidators.ts  input validators returning tagged errors
  browse.ts            the BrowseApi class

Each is now under the 300-line ceiling. `src/api/index.ts` re-exports the
pieces by name instead of star-exporting the old single module, which drops
one `noReExportAll` from that file rather than adding four.

Most of the remaining complexity came from one idiom: building a result by
spreading `...(x === undefined ? {} : { x })` once per optional field, 37
times across the file. `defined()` in src/utils/objects.ts drops
undefined-valued keys in one call, which takes both mappers back under the
cognitive-complexity budget and reads better.

The unit tests split the same way, into browse.test.ts for the class and
browseMappers.test.ts for the pure functions, with the search fixture and
the error patterns hoisted to module scope. Two tests exercising
mapItemSummary were sitting under a `searchActiveItems` describe and moved
across. Splitting the availability picker's nested ternary into an explicit
branch exposed a path the tests never took, since eBay returns
estimatedAvailabilities as either a single-element array or a bare object;
the bare-object case now has a test.

Prose in the files this branch adds also loses its em dashes.

Behaviour is unchanged: 1137 unit and 68 integration tests pass on Node 20,
22 and 24, coverage on the added lines stays at 100%, and the repo-wide
warning count returns to main's 671.
