+++ tests/unit/api/browse.test.ts
@@ -1,482 +1,210 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { Effect } from 'effect';
-import {
-  BrowseApi,
-  buildBrowseFilter,
-  mapItemDetailsResponse,
-  mapItemSummary,
-  mapSearchActiveItemsResponse,
-} from '@/api/other/browse.js';
+import { BrowseApi } from '@/api/other/browse.js';
 import type { EbayApiClient } from '@/api/client.js';
 
-describe('Browse API helpers', () => {
-  describe('buildBrowseFilter', () => {
-    it('returns undefined when no clause applies', () => {
-      expect(buildBrowseFilter({})).toBeUndefined();
+const ERR_LIMIT_MUST_BE_BETWEEN = /limit must be between 1 and 200/;
+const ERR_SORT_MUST_BE_ONE = /sort must be one of/;
+const ERR_OFFSET_MUST_BE_ZERO = /offset must be zero or a multiple of limit \(3\); got 20/;
+const ERR_OFFSET_MUST_BE_BETWEEN = /offset must be between 0 and 10000/;
+const ERR_CONDITIONS_MUST_BE_AN = /conditions must be an array of non-empty strings/;
+const ERR_BUYINGOPTIONS_MUST_BE_AN = /buyingOptions must be an array of non-empty strings/;
+const ERR_PRICEMIN_MUST_NOT_EXCEED = /priceMin \(50\) must not exceed priceMax \(10\)/;
+const ERR_FILTER_ALREADY_CONTAINS_A = /filter already contains a price clause/;
+const ERR_NO_ITEM_FOUND = /No item found/;
+
+let api: BrowseApi;
+let mockClient: EbayApiClient;
+
+beforeEach(() => {
+  mockClient = {
+    get: vi.fn(),
+  } as unknown as EbayApiClient;
+
+  api = new BrowseApi(mockClient);
+});
+
+afterEach(() => {
+  vi.clearAllMocks();
+});
+
+describe('searchActiveItems: request shape', () => {
+  it('calls item_summary/search and maps the response', async () => {
+    vi.mocked(mockClient.get).mockResolvedValue({
+      total: 2,
+      itemSummaries: [{ itemId: 'v1|1|0', title: 'One' }],
     });
 
-    it('builds condition and buying option clauses', () => {
-      expect(buildBrowseFilter({ conditions: ['NEW', 'USED'] })).toBe('conditions:{NEW|USED}');
-      expect(buildBrowseFilter({ buyingOptions: ['AUCTION'] })).toBe('buyingOptions:{AUCTION}');
-    });
+    const result = await Effect.runPromise(
+      api.searchActiveItems({ query: 'camera', limit: 5, offset: 10 }),
+    );
 
-    it('builds price ranges with a default USD currency', () => {
-      expect(buildBrowseFilter({ priceMin: 10, priceMax: 50 })).toBe(
-        'price:[10..50],priceCurrency:USD',
-      );
-      expect(buildBrowseFilter({ priceMin: 10 })).toBe('price:[10..],priceCurrency:USD');
-      expect(buildBrowseFilter({ priceMax: 50, priceCurrency: 'EUR' })).toBe(
-        'price:[..50],priceCurrency:EUR',
-      );
+    expect(mockClient.get).toHaveBeenCalledWith('/buy/browse/v1/item_summary/search', {
+      q: 'camera',
+      limit: 5,
+      offset: 10,
     });
-
-    it('appends the raw filter passthrough last', () => {
-      expect(
-        buildBrowseFilter({
-          conditions: ['NEW'],
-          filter: 'sellers:{user1|user2}',
-        }),
-      ).toBe('conditions:{NEW},sellers:{user1|user2}');
-    });
+    expect(result.items).toEqual([{ itemId: 'v1|1|0', title: 'One' }]);
+    expect(result.total).toBe(2);
   });
+  it('defaults limit to 20 and offset to 0', async () => {
+    vi.mocked(mockClient.get).mockResolvedValue({});
 
-  describe('mapItemSummary', () => {
-    it('maps a full summary', () => {
-      const mapped = mapItemSummary({
-        itemId: 'v1|123|0',
-        title: 'Camera',
-        price: { value: '99.99', currency: 'USD' },
-        condition: 'New',
-        buyingOptions: ['FIXED_PRICE'],
-        itemWebUrl: 'https://www.ebay.com/itm/123',
-        image: { imageUrl: 'https://i.ebayimg.com/img.jpg' },
-        seller: { username: 'cam_seller', feedbackPercentage: '99.8' },
-        shippingOptions: [{ shippingCost: { value: '0.00', currency: 'USD' } }],
-        itemEndDate: '2026-09-01T00:00:00.000Z',
-        itemLocation: { country: 'US' },
-      });
+    const result = await Effect.runPromise(api.searchActiveItems({ query: 'lens' }));
 
-      expect(mapped).toEqual({
-        itemId: 'v1|123|0',
-        title: 'Camera',
-        price: { currency: 'USD', value: '99.99' },
-        condition: 'New',
-        buyingOptions: ['FIXED_PRICE'],
-        itemWebUrl: 'https://www.ebay.com/itm/123',
-        imageUrl: 'https://i.ebayimg.com/img.jpg',
-        seller: 'cam_seller',
-        sellerFeedbackPercentage: '99.8',
-        shippingCost: { currency: 'USD', value: '0.00' },
-        itemEndDate: '2026-09-01T00:00:00.000Z',
-        itemLocationCountry: 'US',
-      });
-    });
-
-    it('returns undefined when itemId or title are missing', () => {
-      expect(mapItemSummary({ title: 'No id' })).toBeUndefined();
-      expect(mapItemSummary({ itemId: 'v1|1|0' })).toBeUndefined();
-      expect(mapItemSummary('not-an-object')).toBeUndefined();
-    });
-
-    it('omits optional fields that are absent or malformed', () => {
-      const mapped = mapItemSummary({
-        itemId: 'v1|1|0',
-        title: 'Sparse',
-        price: { value: 42 },
-        seller: 'not-an-object',
-      });
-
-      expect(mapped).toEqual({ itemId: 'v1|1|0', title: 'Sparse' });
-    });
+    expect(mockClient.get).toHaveBeenCalledWith(
+      '/buy/browse/v1/item_summary/search',
+      expect.objectContaining({ q: 'lens', limit: 20, offset: 0 }),
+    );
+    expect(result.limit).toBe(20);
+    expect(result.offset).toBe(0);
   });
+  it('passes sort, category ids, and the combined filter expression', async () => {
+    vi.mocked(mockClient.get).mockResolvedValue({});
 
-  describe('mapSearchActiveItemsResponse (realistic payload)', () => {
-    // Shaped after eBay's documented item_summary/search response rather than
-    // this module's own assumptions: nested containers, the marketplace/legacy
-    // id fields we deliberately drop, and a thumbnail array we ignore.
-    const EBAY_SEARCH_PAYLOAD = {
-      href: 'https://api.ebay.com/buy/browse/v1/item_summary/search?q=drone&limit=2&offset=0',
-      total: 12_345,
-      next: 'https://api.ebay.com/buy/browse/v1/item_summary/search?q=drone&limit=2&offset=2',
-      limit: 2,
-      offset: 0,
-      itemSummaries: [
-        {
-          itemId: 'v1|254188828753|0',
-          title: 'Drone Quadcopter 4K Camera',
-          leafCategoryIds: ['179697'],
-          categories: [{ categoryId: '179697', categoryName: 'Quadcopters & Multirotors' }],
-          image: { imageUrl: 'https://i.ebayimg.com/images/g/abc/s-l225.jpg' },
-          price: { value: '129.99', currency: 'USD' },
-          itemHref: 'https://api.ebay.com/buy/browse/v1/item/v1%7C254188828753%7C0',
-          seller: {
-            username: 'drone_outlet',
-            feedbackPercentage: '98.7',
-            feedbackScore: 15_432,
-          },
-          condition: 'New',
-          conditionId: '1000',
-          thumbnailImages: [{ imageUrl: 'https://i.ebayimg.com/images/g/abc/s-l64.jpg' }],
-          shippingOptions: [
-            {
-              shippingCostType: 'FIXED',
-              shippingCost: { value: '0.00', currency: 'USD' },
-            },
-          ],
-          buyingOptions: ['FIXED_PRICE'],
-          itemWebUrl: 'https://www.ebay.com/itm/254188828753',
-          itemLocation: { postalCode: '112**', country: 'US' },
-          adultOnly: false,
-          legacyItemId: '254188828753',
-          availableCoupons: false,
-          itemCreationDate: '2026-08-01T10:00:00.000Z',
-          topRatedBuyingExperience: true,
-          priorityListing: false,
-          listingMarketplaceId: 'EBAY_US',
-        },
-      ],
-    };
+    await Effect.runPromise(
+      api.searchActiveItems({
+        query: 'gpu',
+        sort: 'price',
+        categoryIds: '27386',
+        conditions: ['NEW'],
+        priceMax: 500,
+      }),
+    );
 
-    it('maps eBay-shaped payloads and ignores unmapped containers', () => {
-      const result = mapSearchActiveItemsResponse(EBAY_SEARCH_PAYLOAD, {
-        query: 'drone',
-        offset: 0,
-        limit: 2,
-      });
-
-      expect(result.total).toBe(12_345);
-      expect(result.items).toHaveLength(1);
-      expect(result.items[0]).toEqual({
-        itemId: 'v1|254188828753|0',
-        title: 'Drone Quadcopter 4K Camera',
-        price: { currency: 'USD', value: '129.99' },
-        condition: 'New',
-        buyingOptions: ['FIXED_PRICE'],
-        itemWebUrl: 'https://www.ebay.com/itm/254188828753',
-        imageUrl: 'https://i.ebayimg.com/images/g/abc/s-l225.jpg',
-        seller: 'drone_outlet',
-        sellerFeedbackPercentage: '98.7',
-        shippingCost: { currency: 'USD', value: '0.00' },
-        itemLocationCountry: 'US',
-      });
-    });
-
-    it('prefers the pagination eBay returned over the requested window', () => {
-      const result = mapSearchActiveItemsResponse(
-        { ...EBAY_SEARCH_PAYLOAD, offset: 40, limit: 10 },
-        { query: 'drone', offset: 9999, limit: 200 },
-      );
-
-      expect(result.offset).toBe(40);
-      expect(result.limit).toBe(10);
-    });
-
-    it('falls back to the requested window when the payload omits it', () => {
-      const result = mapSearchActiveItemsResponse(
-        { itemSummaries: [] },
-        { query: 'drone', offset: 5, limit: 25 },
-      );
-
-      expect(result.offset).toBe(5);
-      expect(result.limit).toBe(25);
-    });
-  });
-
-  describe('mapSearchActiveItemsResponse', () => {
-    it('maps items and total', () => {
-      const result = mapSearchActiveItemsResponse(
-        {
-          total: 1234,
-          itemSummaries: [{ itemId: 'v1|1|0', title: 'One' }, { title: 'dropped - no id' }],
-        },
-        { query: 'widget', offset: 0, limit: 20 },
-      );
-
-      expect(result.total).toBe(1234);
-      expect(result.items).toEqual([{ itemId: 'v1|1|0', title: 'One' }]);
-      expect(result.query).toBe('widget');
-      expect(result.offset).toBe(0);
-      expect(result.limit).toBe(20);
-    });
-
-    it('returns empty items for unexpected payloads', () => {
-      const context = { query: 'x', offset: 0, limit: 20 };
-      expect(mapSearchActiveItemsResponse(undefined, context)).toEqual({ items: [], ...context });
-      expect(mapSearchActiveItemsResponse({}, context)).toEqual({ items: [], ...context });
-    });
-  });
-
-  describe('mapItemDetailsResponse', () => {
-    it('maps a full item detail', () => {
-      const mapped = mapItemDetailsResponse({
-        itemId: 'v1|123|0',
-        title: 'Camera',
-        price: { value: '99.99', currency: 'USD' },
-        condition: 'New',
-        conditionDescription: 'Open box',
-        shortDescription: 'A nice camera',
-        categoryPath: 'Cameras & Photo|Digital Cameras',
-        buyingOptions: ['FIXED_PRICE', 'BEST_OFFER'],
-        itemWebUrl: 'https://www.ebay.com/itm/123',
-        image: { imageUrl: 'https://i.ebayimg.com/img.jpg' },
-        additionalImages: [{ imageUrl: 'a' }, { imageUrl: 'b' }],
-        seller: { username: 'cam_seller', feedbackPercentage: '99.8' },
-        estimatedAvailabilities: [{ estimatedAvailableQuantity: 3 }],
-        itemEndDate: '2026-09-01T00:00:00.000Z',
-        itemLocation: { city: 'Austin', stateOrProvince: 'TX', country: 'US' },
-        returnTerms: { returnsAccepted: true },
-      });
-
-      expect(mapped).toMatchObject({
-        itemId: 'v1|123|0',
-        title: 'Camera',
-        price: { currency: 'USD', value: '99.99' },
-        conditionDescription: 'Open box',
-        shortDescription: 'A nice camera',
-        additionalImageCount: 2,
-        estimatedAvailableQuantity: 3,
-        itemLocation: 'Austin, TX, US',
-        returnsAccepted: true,
-      });
-    });
-
-    it('returns undefined for payloads without itemId/title', () => {
-      expect(mapItemDetailsResponse(undefined)).toBeUndefined();
-      expect(mapItemDetailsResponse({})).toBeUndefined();
-      expect(mapItemDetailsResponse({ itemId: 'v1|1|0' })).toBeUndefined();
-    });
+    expect(mockClient.get).toHaveBeenCalledWith(
+      '/buy/browse/v1/item_summary/search',
+      expect.objectContaining({
+        sort: 'price',
+        category_ids: '27386',
+        filter: 'conditions:{NEW},price:[..500],priceCurrency:USD',
+      }),
+    );
   });
 });
 
-describe('BrowseApi', () => {
-  let api: BrowseApi;
-  let mockClient: EbayApiClient;
-
-  beforeEach(() => {
-    mockClient = {
-      get: vi.fn(),
-    } as unknown as EbayApiClient;
-
-    api = new BrowseApi(mockClient);
+describe('searchActiveItems: pagination guards', () => {
+  it('rejects a limit above 200', async () => {
+    await expect(
+      Effect.runPromise(api.searchActiveItems({ query: 'x', limit: 201 })),
+    ).rejects.toThrow(ERR_LIMIT_MUST_BE_BETWEEN);
+    expect(mockClient.get).not.toHaveBeenCalled();
   });
-
-  afterEach(() => {
-    vi.clearAllMocks();
+  it('rejects an offset that is not a whole number of pages', async () => {
+    await expect(
+      Effect.runPromise(api.searchActiveItems({ query: 'x', limit: 3, offset: 20 })),
+    ).rejects.toThrow(ERR_OFFSET_MUST_BE_ZERO);
+    expect(mockClient.get).not.toHaveBeenCalled();
   });
+  it('accepts an offset that is a multiple of limit', async () => {
+    vi.mocked(mockClient.get).mockResolvedValue({});
 
-  describe('searchActiveItems', () => {
-    it('calls item_summary/search and maps the response', async () => {
-      vi.mocked(mockClient.get).mockResolvedValue({
-        total: 2,
-        itemSummaries: [{ itemId: 'v1|1|0', title: 'One' }],
-      });
+    await Effect.runPromise(api.searchActiveItems({ query: 'x', limit: 3, offset: 21 }));
 
-      const result = await Effect.runPromise(
-        api.searchActiveItems({ query: 'camera', limit: 5, offset: 10 }),
-      );
+    expect(mockClient.get).toHaveBeenCalledWith(
+      '/buy/browse/v1/item_summary/search',
+      expect.objectContaining({ limit: 3, offset: 21 }),
+    );
+  });
+  it('rejects an offset above the Browse maximum', async () => {
+    await expect(
+      Effect.runPromise(api.searchActiveItems({ query: 'x', limit: 1, offset: 10_001 })),
+    ).rejects.toThrow(ERR_OFFSET_MUST_BE_BETWEEN);
+    expect(mockClient.get).not.toHaveBeenCalled();
+  });
+  it('accepts an offset at the Browse maximum', async () => {
+    vi.mocked(mockClient.get).mockResolvedValue({});
 
-      expect(mockClient.get).toHaveBeenCalledWith('/buy/browse/v1/item_summary/search', {
-        q: 'camera',
-        limit: 5,
-        offset: 10,
-      });
-      expect(result.items).toEqual([{ itemId: 'v1|1|0', title: 'One' }]);
-      expect(result.total).toBe(2);
-    });
+    await Effect.runPromise(api.searchActiveItems({ query: 'x', offset: 10_000 }));
 
-    it('defaults limit to 20 and offset to 0', async () => {
-      vi.mocked(mockClient.get).mockResolvedValue({});
+    expect(mockClient.get).toHaveBeenCalledWith(
+      '/buy/browse/v1/item_summary/search',
+      expect.objectContaining({ offset: 10_000 }),
+    );
+  });
+});
 
-      const result = await Effect.runPromise(api.searchActiveItems({ query: 'lens' }));
+describe('searchActiveItems: input validation', () => {
+  it('rejects an unsupported sort value', async () => {
+    await expect(
+      Effect.runPromise(api.searchActiveItems({ query: 'x', sort: 'distance' as never })),
+    ).rejects.toThrow(ERR_SORT_MUST_BE_ONE);
+    expect(mockClient.get).not.toHaveBeenCalled();
+  });
+  it('rejects non-string entries in conditions and buyingOptions', async () => {
+    await expect(
+      Effect.runPromise(api.searchActiveItems({ query: 'x', conditions: [1 as never] })),
+    ).rejects.toThrow(ERR_CONDITIONS_MUST_BE_AN);
+    await expect(
+      Effect.runPromise(api.searchActiveItems({ query: 'x', buyingOptions: ['' as never] })),
+    ).rejects.toThrow(ERR_BUYINGOPTIONS_MUST_BE_AN);
+    expect(mockClient.get).not.toHaveBeenCalled();
+  });
+  it('rejects a negative price bound', async () => {
+    await expect(
+      Effect.runPromise(api.searchActiveItems({ query: 'x', priceMin: -1 })),
+    ).rejects.toThrow();
+    expect(mockClient.get).not.toHaveBeenCalled();
+  });
+  it('rejects an inverted price window', async () => {
+    await expect(
+      Effect.runPromise(api.searchActiveItems({ query: 'x', priceMin: 50, priceMax: 10 })),
+    ).rejects.toThrow(ERR_PRICEMIN_MUST_NOT_EXCEED);
+    expect(mockClient.get).not.toHaveBeenCalled();
+  });
+  it('rejects a missing query', async () => {
+    await expect(Effect.runPromise(api.searchActiveItems({ query: '' }))).rejects.toThrow();
+    expect(mockClient.get).not.toHaveBeenCalled();
+  });
+});
 
-      expect(mockClient.get).toHaveBeenCalledWith(
-        '/buy/browse/v1/item_summary/search',
-        expect.objectContaining({ q: 'lens', limit: 20, offset: 0 }),
-      );
-      expect(result.limit).toBe(20);
-      expect(result.offset).toBe(0);
-    });
-
-    it('passes sort, category ids, and the combined filter expression', async () => {
-      vi.mocked(mockClient.get).mockResolvedValue({});
-
-      await Effect.runPromise(
+describe('searchActiveItems: raw filter conflicts', () => {
+  it('rejects a raw filter whose price clause collides with priceMin/priceMax', async () => {
+    await expect(
+      Effect.runPromise(
         api.searchActiveItems({
-          query: 'gpu',
-          sort: 'price',
-          categoryIds: '27386',
-          conditions: ['NEW'],
-          priceMax: 500,
+          query: 'x',
+          priceMax: 100,
+          filter: 'price:[5..10],priceCurrency:USD',
         }),
-      );
+      ),
+    ).rejects.toThrow(ERR_FILTER_ALREADY_CONTAINS_A);
+    expect(mockClient.get).not.toHaveBeenCalled();
+  });
+  it('allows a raw filter without a price clause alongside price bounds', async () => {
+    vi.mocked(mockClient.get).mockResolvedValue({});
 
-      expect(mockClient.get).toHaveBeenCalledWith(
-        '/buy/browse/v1/item_summary/search',
-        expect.objectContaining({
-          sort: 'price',
-          category_ids: '27386',
-          filter: 'conditions:{NEW},price:[..500],priceCurrency:USD',
-        }),
-      );
+    await Effect.runPromise(
+      api.searchActiveItems({ query: 'x', priceMax: 100, filter: 'sellers:{acme}' }),
+    );
+
+    expect(mockClient.get).toHaveBeenCalledWith(
+      '/buy/browse/v1/item_summary/search',
+      expect.objectContaining({
+        filter: 'price:[..100],priceCurrency:USD,sellers:{acme}',
+      }),
+    );
+  });
+});
+
+describe('getItemDetails', () => {
+  it('URL-encodes the item id and maps the response', async () => {
+    vi.mocked(mockClient.get).mockResolvedValue({
+      itemId: 'v1|110587051479|0',
+      title: 'Camera',
     });
 
-    it('rejects a limit above 200', async () => {
-      await expect(
-        Effect.runPromise(api.searchActiveItems({ query: 'x', limit: 201 })),
-      ).rejects.toThrow(/limit must be between 1 and 200/);
-      expect(mockClient.get).not.toHaveBeenCalled();
-    });
+    const result = await Effect.runPromise(api.getItemDetails({ itemId: 'v1|110587051479|0' }));
 
-    it('rejects an unsupported sort value', async () => {
-      await expect(
-        Effect.runPromise(api.searchActiveItems({ query: 'x', sort: 'distance' as never })),
-      ).rejects.toThrow(/sort must be one of/);
-      expect(mockClient.get).not.toHaveBeenCalled();
-    });
-
-    it('maps an auction summary using currentBidPrice and bidCount', () => {
-      const mapped = mapItemSummary({
-        itemId: 'v1|999|0',
-        title: 'Vintage SLR, no reserve',
-        currentBidPrice: { value: '42.50', currency: 'USD' },
-        bidCount: 7,
-        buyingOptions: ['AUCTION'],
-        itemEndDate: '2026-09-01T18:00:00.000Z',
-      });
-
-      expect(mapped).toMatchObject({
-        price: { currency: 'USD', value: '42.50' },
-        bidCount: 7,
-        buyingOptions: ['AUCTION'],
-        itemEndDate: '2026-09-01T18:00:00.000Z',
-      });
-    });
-
-    it('prefers a fixed price over currentBidPrice when both are present', () => {
-      const mapped = mapItemSummary({
-        itemId: 'v1|998|0',
-        title: 'Auction with Buy It Now',
-        price: { value: '99.00', currency: 'USD' },
-        currentBidPrice: { value: '42.50', currency: 'USD' },
-        buyingOptions: ['AUCTION', 'FIXED_PRICE'],
-      });
-
-      expect(mapped?.price).toEqual({ currency: 'USD', value: '99.00' });
-    });
-
-    it('rejects an offset that is not a whole number of pages', async () => {
-      await expect(
-        Effect.runPromise(api.searchActiveItems({ query: 'x', limit: 3, offset: 20 })),
-      ).rejects.toThrow(/offset must be zero or a multiple of limit \(3\); got 20/);
-      expect(mockClient.get).not.toHaveBeenCalled();
-    });
-
-    it('accepts an offset that is a multiple of limit', async () => {
-      vi.mocked(mockClient.get).mockResolvedValue({});
-
-      await Effect.runPromise(api.searchActiveItems({ query: 'x', limit: 3, offset: 21 }));
-
-      expect(mockClient.get).toHaveBeenCalledWith(
-        '/buy/browse/v1/item_summary/search',
-        expect.objectContaining({ limit: 3, offset: 21 }),
-      );
-    });
-
-    it('rejects an offset above the Browse maximum', async () => {
-      await expect(
-        Effect.runPromise(api.searchActiveItems({ query: 'x', limit: 1, offset: 10_001 })),
-      ).rejects.toThrow(/offset must be between 0 and 10000/);
-      expect(mockClient.get).not.toHaveBeenCalled();
-    });
-
-    it('accepts an offset at the Browse maximum', async () => {
-      vi.mocked(mockClient.get).mockResolvedValue({});
-
-      await Effect.runPromise(api.searchActiveItems({ query: 'x', offset: 10_000 }));
-
-      expect(mockClient.get).toHaveBeenCalledWith(
-        '/buy/browse/v1/item_summary/search',
-        expect.objectContaining({ offset: 10_000 }),
-      );
-    });
-
-    it('rejects non-string entries in conditions and buyingOptions', async () => {
-      await expect(
-        Effect.runPromise(api.searchActiveItems({ query: 'x', conditions: [1 as never] })),
-      ).rejects.toThrow(/conditions must be an array of non-empty strings/);
-      await expect(
-        Effect.runPromise(api.searchActiveItems({ query: 'x', buyingOptions: ['' as never] })),
-      ).rejects.toThrow(/buyingOptions must be an array of non-empty strings/);
-      expect(mockClient.get).not.toHaveBeenCalled();
-    });
-
-    it('rejects a negative price bound', async () => {
-      await expect(
-        Effect.runPromise(api.searchActiveItems({ query: 'x', priceMin: -1 })),
-      ).rejects.toThrow();
-      expect(mockClient.get).not.toHaveBeenCalled();
-    });
-
-    it('rejects an inverted price window', async () => {
-      await expect(
-        Effect.runPromise(api.searchActiveItems({ query: 'x', priceMin: 50, priceMax: 10 })),
-      ).rejects.toThrow(/priceMin \(50\) must not exceed priceMax \(10\)/);
-      expect(mockClient.get).not.toHaveBeenCalled();
-    });
-
-    it('rejects a raw filter whose price clause collides with priceMin/priceMax', async () => {
-      await expect(
-        Effect.runPromise(
-          api.searchActiveItems({
-            query: 'x',
-            priceMax: 100,
-            filter: 'price:[5..10],priceCurrency:USD',
-          }),
-        ),
-      ).rejects.toThrow(/filter already contains a price clause/);
-      expect(mockClient.get).not.toHaveBeenCalled();
-    });
-
-    it('allows a raw filter without a price clause alongside price bounds', async () => {
-      vi.mocked(mockClient.get).mockResolvedValue({});
-
-      await Effect.runPromise(
-        api.searchActiveItems({ query: 'x', priceMax: 100, filter: 'sellers:{acme}' }),
-      );
-
-      expect(mockClient.get).toHaveBeenCalledWith(
-        '/buy/browse/v1/item_summary/search',
-        expect.objectContaining({
-          filter: 'price:[..100],priceCurrency:USD,sellers:{acme}',
-        }),
-      );
-    });
-
-    it('rejects a missing query', async () => {
-      await expect(Effect.runPromise(api.searchActiveItems({ query: '' }))).rejects.toThrow();
-      expect(mockClient.get).not.toHaveBeenCalled();
-    });
+    expect(mockClient.get).toHaveBeenCalledWith('/buy/browse/v1/item/v1%7C110587051479%7C0');
+    expect(result).toEqual({ itemId: 'v1|110587051479|0', title: 'Camera' });
   });
 
-  describe('getItemDetails', () => {
-    it('URL-encodes the item id and maps the response', async () => {
-      vi.mocked(mockClient.get).mockResolvedValue({
-        itemId: 'v1|110587051479|0',
-        title: 'Camera',
-      });
+  it('fails with a tagged input error when the response has no item', async () => {
+    vi.mocked(mockClient.get).mockResolvedValue({});
 
-      const result = await Effect.runPromise(api.getItemDetails({ itemId: 'v1|110587051479|0' }));
-
-      expect(mockClient.get).toHaveBeenCalledWith('/buy/browse/v1/item/v1%7C110587051479%7C0');
-      expect(result).toEqual({ itemId: 'v1|110587051479|0', title: 'Camera' });
-    });
-
-    it('fails with a tagged input error when the response has no item', async () => {
-      vi.mocked(mockClient.get).mockResolvedValue({});
-
-      await expect(Effect.runPromise(api.getItemDetails({ itemId: 'v1|1|0' }))).rejects.toThrow(
-        /No item found/,
-      );
-    });
+    await expect(Effect.runPromise(api.getItemDetails({ itemId: 'v1|1|0' }))).rejects.toThrow(
+      ERR_NO_ITEM_FOUND,
+    );
   });
 });
